+++ API/controllers/user.go
@@ -8,16 +8,22 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
+// GetUserById handle /user/id (GET)
 func GetUserById(c *gin.Context) {
 	c.JSON(http.StatusOK, repository.GetUserById(c.Params.ByName("id")))
 }
 
+// PostUser handle /user for creating a new user (POST)
 func PostUser(c *gin.Context) {
-
 	// Validate input
 	var input models.PostUser
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if repository.GetUserByEmail(input.Email).UUID != "" {
+		c.JSON(http.StatusConflict, gin.H{"error": "User with this mail already exist"})
 		return
 	}
 
@@ -26,12 +32,17 @@ func PostUser(c *gin.Context) {
 	c.JSON(http.StatusCreated, input)
 }
 
+// PutUserById handle /user/id for editing an existing user
 func PutUserById(c *gin.Context) {
-
 	// Validate input
 	var input models.PostUser
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if repository.GetUserByEmail(input.Email).UUID != "" {
+		c.JSON(http.StatusConflict, gin.H{"error": "This email is already taken by another user"})
 		return
 	}
 
@@ -40,7 +51,13 @@ func PutUserById(c *gin.Context) {
 	c.JSON(http.StatusOK, input)
 }
 
+// DeleteUserById handle /user/id for deleting an existing user
 func DeleteUserById(c *gin.Context) {
+	if repository.GetUserById(c.Params.ByName("id")).UUID != "" {
+		c.JSON(http.StatusConflict, gin.H{"err": "User with this ID does not exist"})
+		return
+	}
+
 	repository.DeleteUserById(c.Params.ByName("id"))
 
 	c.String(http.StatusOK, "User successfully deleted")
